backend/requirements.txt
# Security & Rate Limiting
slowapi==0.1.9
limits==5.0.0

# Core Dependencies
aiofiles==25.1.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.14.0
passlib==1.7.4
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.1.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.7.1
pydantic_core==2.41.5
PyPDF2==3.0.1
pypdfium2==5.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0

# PostgreSQL driver
psycopg2-binary==2.9.10

# Export functionality
reportlab>=4.0
openpyxl>=3.1

# Email
aiosmtplib>=2.0

# 2FA
pyotp>=2.9
